requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

function-cooldowns==1.6.0
nextcord==2.3.2
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
pymongo==4.3.3
